% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovTamal.R
\name{clx}
\alias{clx}
\title{sandwich in groups}
\usage{
clx(fm = et, dfcw = dfcw, cluster = gx)
}
\arguments{
\item{fm}{estimate of transformed data}

\item{dfcw}{robust degrees of freedom}

\item{cluster}{group cluster}
}
\description{
XXXX
}
