% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.transform.R
\name{panel.transform}
\alias{panel.transform}
\title{Transform Panel Data}
\usage{
panel.transform(formula, data, model = "within", index = c("id", "year"))
}
\arguments{
\item{formula}{a formula, indicating the covariates.}

\item{data}{a data frame}

\item{model}{a model, "fd" or "within"}

\item{index}{id and group structure, e.g. c("id","year")}
}
\value{
data.frame
}
\description{
panel.transform converts the a panel data frame to time-demeand data or to first differenced data.
}
\details{
convert data frame into transformed data
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
