% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plm2lm.R
\name{plm2lm}
\alias{plm2lm}
\title{Transform plm to lm object}
\usage{
plm2lm(estimate, reestimate = T)
}
\arguments{
\item{estimate}{an object of class \code{"plm"}.}

\item{reestimate}{logical, if TRUE the panel data model is reestimated with a linear regression (\code{lm}).}
}
\value{
A \code{lm} object
}
\description{
plm2lm, converts the a \code{plm} object to an \code{lm} object.
The information saved in a \code{plm} object is used to obtain an transformed dataset
(e.g. time-demeaned or first differences)
in order to restimate an OLS model with \code{lm}.
}
\details{
estimate - The \code{plm} object estimated by one of the methods \code{within}, \code{fd}.
reestimate - If FALSE, an error displays - A conversion without reestimation is currently not implmented.
}
\note{
Reestimation of results leads to an increase in processing requairemets.
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
