% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcovTamal.R
\name{vcovTamal}
\alias{vcovTamal}
\title{Group clustering of panel data esitmates}
\usage{
vcovTamal(estimate, data, groupvar)
}
\arguments{
\item{estimate}{an object of class \code{"plm"}.}

\item{data}{the data frame object used to create \code{estimate} object. Can be a data.frame or a pdata.frame object}

\item{groupvar}{a string indicating a column in \code{data} to indexes the group structure.}
}
\value{
A matrix containing the covariance matrix estimate
}
\description{
vcovTamal, creates a robust covariance variance matrix clustered at group level for \code{lm} objects.
Clustering is based on Angist and Pischke (XXXX), XXXX and XXX resulting in Stata(TM) like standard errors of
regression coefficients.
}
\details{
Clustering \cite{plm} objects estimated by a "within" model: \cr
Details forthcomming... \cr

Clustering \cite{plm} objects estimated by a "fd" model: \cr
Details forthcomming... \cr
}
\note{
The function uses the \code{estimate} and \code{data} to construct a transfomed data frame
- first differenced or time-demeaned. The esitmate is then re-estimated by OLS with \code{lm} in order to extract the
covarinace matrix and correct for degrees of freedom and use a sandwich estimate with the group structure.
}
\examples{
##packages
library(foreign)
library(clubTamal)
library(lmtest)
library(plm)
library(clubSandwich)
library(spd4testing)

##data
d<-spd4testing()
d

#manipulate data
# d<-pdata.frame(d)

##formula
f<- formula(y ~ x + factor(year))

##standard estiatmion
e<-plm(formula=f,data=d,model="fd")
summary(e)
e<-plm(formula=f,data=d,model="within")
summary(e)

##clustering
#no clustering
v<-e$vcov
coeftest(e)
#clustering at id level with plm-package
v<-vcovHC(e,type="HC1",cluster="group",tol=1*10^-20 )
coeftest(e,v)

##clustering at group level with clubSandwich-package
v<-vcovCR(e,d$gid,type="CR1")
coef_test(e,vcov=v,test="naive-t")
coeftest(e,v)

##clustering at group level with clubTamal
v<-vcovTamal(estimate=e,data=d,groupvar="gid")
coeftest(e,v)


##clustering at group level with STATA (TM)
#export data
tdir<-tempdir()
write.csv(d,file.path(tdir,"spd.csv"))
#write stata file
sink(file.path(tdir,"clusterStata.do"))
cat(paste0('insheet using "',file.path("spd.csv"),'", clear'))
cat("\\n")
cat('xtset id year')
cat("\\n")
cat('xtreg y x i.year, fe')
cat("\\n")
cat('xtreg y x i.year, fe vce(cluster gid)')
cat("\\n")
cat('reg d.y d.x i.year')
cat("\\n")
cat('reg d.y d.x i.year, vce(cluster gid)')
sink()
## RUN Stata file

}
\references{
Angist and Pischke (XXXX)
}
\author{
Elías Cisneros <ec@elias-cisneros.de>
}
